Move custom http client to a separate file
